+++ R/filterCDS.R
@@ -8,15 +8,17 @@
 #' @import GenomicRanges
 #' @importFrom methods as is
 #' @examples
-#' library(GenomicFeatures)
+#' #library(GenomicFeatures)
 #' library(BSgenome.Drerio.UCSC.danRer10)
-#' txdb <- makeTxDbFromGFF(system.file("extdata",
-#'           "Danio_rerio.GRCz10.91.chr1.gtf.gz",
-#'           package="ribosomeProfilingQC"),
-#'           organism = "Danio rerio",
-#'           chrominfo = seqinfo(Drerio)["chr1"],
-#'           taxonomyId = 7955)
-#' CDS <- prepareCDS(txdb)
+#' #txdb <- makeTxDbFromGFF(system.file("extdata",
+#'  #         "Danio_rerio.GRCz10.91.chr1.gtf.gz",
+#'  #         package="ribosomeProfilingQC"),
+#'  #         organism = "Danio rerio",
+#'  #         chrominfo = seqinfo(Drerio)["chr1"],
+#'  #         taxonomyId = 7955)
+#' #CDS <- prepareCDS(txdb)
+#' CDS <- readRDS(system.file("extdata", "CDS.rds",
+#'                            package="ribosomeProfilingQC"))
 #' filterCDS(CDS)
 filterCDS <- function(CDS, sizeCutoff = 100L){
   sizeCutoff <- sizeCutoff[1]