+++ R/getFPKM.R
@@ -1,5 +1,6 @@
 #' get FPKM values for counts
-#' @description Calculate Fragments Per Kilobase of transcript per Million mapped reads (FPKM) for counts.
+#' @description Calculate Fragments Per Kilobase of transcript per
+#' Million mapped reads (FPKM) for counts.
 #' @param counts output of \link{countReads} or \link{normByRUVs}
 #' @param gtf GTF file name for annotation.
 #' @param level transcript or gene level.