+++ R/ggBar.R
@@ -33,7 +33,8 @@ ggBar <- function(height, fill="gray80", draw=TRUE, xlab, ylab, postfix){
     geom_bar(stat = "identity", fill=fill) +
     theme_classic()
   if(!missing(postfix)) suppressWarnings(
-    plot <- plot + geom_bar_text(aes(label=label), grow = FALSE, contrast = TRUE))
+    plot <- plot +
+      geom_bar_text(aes(label=label), grow = FALSE, contrast = TRUE))
   if(!missing(xlab)) plot <- plot+xlab(xlab)
   if(!missing(ylab)) plot <- plot+ylab(ylab)
   if(draw) {