+++ R/readsEndPlot.R
@@ -19,19 +19,21 @@
 #'                            package="ribosomeProfilingQC")
 #' yieldSize <- 10000000
 #' bamfile <- BamFile(bamfilename, yieldSize = yieldSize)
-#' library(GenomicFeatures)
+#' #library(GenomicFeatures)
 #' library(BSgenome.Drerio.UCSC.danRer10)
-#' txdb <- makeTxDbFromGFF(system.file("extdata",
-#'           "Danio_rerio.GRCz10.91.chr1.gtf.gz",
-#'           package="ribosomeProfilingQC"),
-#'           organism = "Danio rerio",
-#'           chrominfo = seqinfo(Drerio)["chr1"],
-#'           taxonomyId = 7955)
-#' CDS <- prepareCDS(txdb)
+#' #txdb <- makeTxDbFromGFF(system.file("extdata",
+#'  #         "Danio_rerio.GRCz10.91.chr1.gtf.gz",
+#'  #         package="ribosomeProfilingQC"),
+#'  #         organism = "Danio rerio",
+#'  #         chrominfo = seqinfo(Drerio)["chr1"],
+#'  #         taxonomyId = 7955)
+#' #CDS <- prepareCDS(txdb)
+#' CDS <- readRDS(system.file("extdata", "CDS.rds",
+#'                            package="ribosomeProfilingQC"))
 #' readsEndPlot(bamfile, CDS, toStartCodon=TRUE)
 #' readsEndPlot(bamfile, CDS, toStartCodon=TRUE, fiveEnd=FALSE)
-#' readsEndPlot(bamfile, CDS, toStartCodon=FALSE)
-#' readsEndPlot(bamfile, CDS, toStartCodon=FALSE, fiveEnd=FALSE)
+#' #readsEndPlot(bamfile, CDS, toStartCodon=FALSE)
+#' #readsEndPlot(bamfile, CDS, toStartCodon=FALSE, fiveEnd=FALSE)
 readsEndPlot <- function(bamfile, CDS, toStartCodon=TRUE,
                          fiveEnd=TRUE, window=c(-29, 30),
                          readLen=25:30){