+++ R/ribosomeReleaseScore.R
@@ -1,14 +1,18 @@
 #' Ribosome Release Score (RRS)
-#' @description RRS is calculated as the ratio of translational efficiency in the CDS with RPFs in the 3'UTR.
-#' @param cdsTE,utr3TE translational efficiency of CDS and UTR3 region. Output of \link{translationalEfficiency}
-#' @param CDSsampleOrder,UTR3sampleOrder sample order of cdsTE and utr3TE The parameters
-#' are used to make sure that the order of CDS and UTR3 in TE is corresponding samples.
+#' @description RRS is calculated as the ratio of translational efficiency in
+#'  the CDS with RPFs in the 3'UTR.
+#' @param cdsTE,utr3TE translational efficiency of CDS and UTR3 region.
+#'  Output of \link{translationalEfficiency}
+#' @param CDSsampleOrder,UTR3sampleOrder sample order of cdsTE and utr3TE.
+#'  The parameters are used to make sure that the order of CDS and UTR3 in TE
+#'   is corresponding samples.
 #' @param pseudocount the number will be add to sum of reads count to avoid X/0.
 #' @param log2 Do log2 transform or not.
 #' @return a vector of RRS.
 #' @export
 #' @examples
 #'
+#' \dontrun{
 #' path <- system.file("extdata", package="ribosomeProfilingQC")
 #' RPFs <- dir(path, "RPF.*?\\.[12].bam$", full.names=TRUE)
 #' RNAs <- dir(path, "mRNA.*?\\.[12].bam$", full.names=TRUE)
@@ -18,6 +22,7 @@
 #' TE90 <- translationalEfficiency(cvgs, window = 90)
 #' TE90.utr3 <- translationalEfficiency(cvgs.utr3, window = 90)
 #' rrs <- ribosomeReleaseScore(TE90, TE90.utr3)
+#'}
 #'
 ribosomeReleaseScore <- function(cdsTE, utr3TE, CDSsampleOrder, UTR3sampleOrder,
                                  pseudocount=0, log2=FALSE){
@@ -26,7 +31,8 @@ ribosomeReleaseScore <- function(cdsTE, utr3TE, CDSsampleOrder, UTR3sampleOrder,
   }
   if(!all(c("RPFs", "mRNA", "TE") %in% names(cdsTE)) ||
      !all(c("RPFs", "mRNA", "TE") %in% names(utr3TE))){
-    stop("cdsTE and utr3TE must be output of translationalEfficiency And RPFs, mRNA and TE must be available.")
+    stop("cdsTE and utr3TE must be output of translationalEfficiency And RPFs,
+         mRNA and TE must be available.")
   }
   cdsTE <- cdsTE$TE
   utr3TE <- utr3TE$TE