+++ R/simulateRPF.R
@@ -12,7 +12,7 @@
 #' frame0, frame1 and frame2
 #' @param DEregions The regions with differential reads in exon,
 #' utr5 and utr3.
-#' @param size dispersion parameter. Must be strictly positive, need not be integer.
+#' @param size dispersion parameter. Must be strictly positive.
 #' @param sd standard deviations.
 #' @param minDElevel minimal differential level. default: log2(2).
 #' @param includeReadsSeq logical. Include reads sequence or not.
@@ -104,14 +104,19 @@ simulateRPF <- function(txdb, outPath, genome, samples = 6,
   n <- length(CDS.length.3)
   x <- rnbinom(n, mu=readsPerSample/n, size = size)
   ## normal distribution for samples
-  y <- lapply(x, function(mu) round(rnorm(n=length(samples), mean = mu, sd=sd*mu)))
+  y <- lapply(x, function(mu)
+    round(rnorm(n=length(samples), mean = mu, sd=sd*mu)))
 
   ## assign position
-  p <- mapply(CDS.length.3, y, u5Len[names(CDS.length.3)], fullLen[names(CDS.length.3)],
+  p <- mapply(CDS.length.3, y,
+              u5Len[names(CDS.length.3)],
+              fullLen[names(CDS.length.3)],
               FUN = function(m, n, ups, dws){
     lapply(n, function(.n){
-      .min <- max(0, ceiling((psite-1-ups)/3))## if there is no utr5, it should leave the space for p0 to psite
-      .max <- min(m-1, max(1, floor((dws+psite-1-ups)/3))) ## if there is no utr3, it should leave the space from psite to end
+      ## if there is no utr5, it should leave the space for p0 to psite
+      .min <- max(0, ceiling((psite-1-ups)/3))
+      ## if there is no utr3, it should leave the space from psite to end
+      .max <- min(m-1, max(1, floor((dws+psite-1-ups)/3)))
       if(.min>.max) return(NULL)
       3*round(runif(n=.n,
                     min=.min,
@@ -142,16 +147,19 @@ simulateRPF <- function(txdb, outPath, genome, samples = 6,
     cds_de <- cds[subjectHits(ol_DE)]
     cds_de <- split(cds_de, cds_de$feature)
     ## add de 5'UTR and 3'UTRs
-    ## sample the reads from CDS region and shift it into 5'UTR or 3'UTR for up group
+    ## sample the reads from CDS region and shift it into
+    ## 5'UTR or 3'UTR for up group
     ## all psite is in CDS now
     ## 1. get the length of 5'UTR and 3'UTR > readsLen
     ## 2. split it into two group up and down
-    ## 3. get total reads for each txs, and reditribute the reads in 5'UTR (or 3'UTR) + CDS
+    ## 3. get total reads for each txs,
+    ##    and reditribute the reads in 5'UTR (or 3'UTR) + CDS
     if(length(cds_de$UTR3)>0 || length(cds_de$UTR5)>0){
       de_evt <- c(cds_de$UTR3, cds_de$UTR5)
       execpt <- width(de_evt)< readsLen
       if(any(execpt)){
-        message("There are ", sum(execpt), " DE UTRs width smaller than readsLen.")
+        message("There are ", sum(execpt),
+                " DE UTRs width smaller than readsLen.")
       }
       de_evt <- de_evt[!execpt]
       pp <- unlist(GRangesList(p))
@@ -220,7 +228,8 @@ simulateRPF <- function(txdb, outPath, genome, samples = 6,
     ## move the reads from down group to up group
     if(length(cds_de$CDS)>0){
       de_evt <- cds_de$CDS
-      de_evt_rg <- GRanges(de_evt$tx_name, IRanges(de_evt$internalPos+1, de_evt$wid.cumsum))
+      de_evt_rg <- GRanges(de_evt$tx_name,
+                           IRanges(de_evt$internalPos+1, de_evt$wid.cumsum))
       pp <- unlist(GRangesList(p))
       pp$group <- rep(names(p), lengths(p))
       pp$gp <- pp$group %in% samples[group1]
@@ -228,7 +237,8 @@ simulateRPF <- function(txdb, outPath, genome, samples = 6,
       de_evt_ol <- suppressWarnings(findOverlaps(de_evt_rg, pp))
       execpt <- !seq_along(de_evt_rg) %in% queryHits(de_evt_ol)
       if(any(execpt)){
-        message("There are ", sum(execpt), " DE GRanges does not contain any simulated reads.")
+        message("There are ", sum(execpt),
+                " DE GRanges does not contain any simulated reads.")
       }
       de_evt <- de_evt[!execpt]
       de_evt_rg <- de_evt_rg[!execpt]
@@ -241,12 +251,15 @@ simulateRPF <- function(txdb, outPath, genome, samples = 6,
       de_aliq <- 2*de_cnt/de_aliq
       de_to_move <- round((de_aliq*abs(de_times)-de_cnt)*sign(de_times))
       de_evt_ol <- suppressWarnings(findOverlaps(de_evt_rg, pp))
-      ## split the all reads in the event and to be moved reads to by de_evt index
+      ## split the all reads in the event and
+      ## to be moved reads to by de_evt index
       de_evt_reads <- split(pp[subjectHits(de_evt_ol)], queryHits(de_evt_ol))
       de_to_move <- de_to_move[as.numeric(names(de_evt_reads))]
       notsig <- abs(de_to_move) > lengths(de_evt_reads) | de_to_move == 0
       if(any(notsig)){
-        message("There are ", sum(notsig), " DE GRanges does not contain enough simulated reads to show difference.")
+        message("There are ", sum(notsig),
+                " DE GRanges does not contain enough
+                simulated reads to show difference.")
       }
       renameM <- samples[c(group1, group2)]
       names(renameM) <- samples[c(group2, group1)]
@@ -262,7 +275,8 @@ simulateRPF <- function(txdb, outPath, genome, samples = 6,
                                  x[.reads$gp]
                                }
                                if(length(x)>0){
-                                 .id <- sample(x = x, size = .size, replace = TRUE)
+                                 .id <- sample(x = x, size = .size,
+                                               replace = TRUE)
                                  .id <- unique(.id)
                                  .reads$group[.id] <- renameM[.reads$group[.id]]
                                }
@@ -282,13 +296,14 @@ simulateRPF <- function(txdb, outPath, genome, samples = 6,
   p <- lapply(p, function(.sample){
     .sample <- shift(.sample, shift = u5Len[as.character(seqnames(.sample))])
     start(.sample)[start(.sample)<0] <- 0 ## no need, but keep
-    .sample <- shift(.sample, shift = 1) ## shift 1, why not move to assign position step?
+    ## shift 1, why not move to assign position step?
+    .sample <- shift(.sample, shift = 1)
     .sample[start(.sample)>fullLen[as.character(seqnames(.sample))]] <-
       shift(.sample[start(.sample)>fullLen[as.character(seqnames(.sample))]],
             shift = -3)
     start(.sample[start(.sample)>fullLen[as.character(seqnames(.sample))]]) <-
-      fullLen[as.character(seqnames(.sample))[
-        start(.sample)>fullLen[as.character(seqnames(.sample))]]] ## no need, but keep
+      fullLen[as.character(seqnames(.sample))[## no need, but keep
+        start(.sample)>fullLen[as.character(seqnames(.sample))]]]
     .sample
   })
 
@@ -338,9 +353,11 @@ simulateRPF <- function(txdb, outPath, genome, samples = 6,
                            .ir <- c(ss[id1:id2], es[id1:id2])
                            .ir <- .ir[.ir>ir[1] & .ir<ir[2]]
                            if(std[id1]){
-                             .ir <- c(.ir, es[id2][es[id2] %in% ir], ss[id1][ss[id1] %in% ir])
+                             .ir <- c(.ir, es[id2][es[id2] %in% ir],
+                                      ss[id1][ss[id1] %in% ir])
                            }else{
-                             .ir <- c(.ir, es[id1][es[id1] %in% ir], ss[id2][ss[id2] %in% ir])
+                             .ir <- c(.ir, es[id1][es[id1] %in% ir],
+                                      ss[id2][ss[id2] %in% ir])
                            }
                            sort(c(ir, .ir))
                          }, SIMPLIFY = FALSE)
@@ -356,15 +373,16 @@ simulateRPF <- function(txdb, outPath, genome, samples = 6,
             sep="", collapse = "")
     })
     #njunc <- abs(pos_start$i_id - pos_stop$i_id)
-    reads <- GAlignments(seqnames = seqnames(pos_start),
-                         strand = strand(pos_start),
-                         cigar = cigar,
-                         pos = st,
-                         names = paste(as.character(seqnames(.sample)),
-                                       unlist(lapply(runLength(seqnames(.sample)),
+    reads <-
+      GAlignments(seqnames = seqnames(pos_start),
+                  strand = strand(pos_start),
+                  cigar = cigar,
+                  pos = st,
+                  names = paste(as.character(seqnames(.sample)),
+                                unlist(lapply(runLength(seqnames(.sample)),
                                               seq.int)),
-                                       sep = "_"),
-                         tx = as.character(seqnames(.sample)))
+                                sep = "_"),
+                  tx = as.character(seqnames(.sample)))
     seqinfo(reads) <- seqinfo(txdb)[seqlevels(reads)]
 
     if(!mis_genome){
@@ -394,16 +412,19 @@ simulateRPF <- function(txdb, outPath, genome, samples = 6,
   # re <- start(p[[1]][seqnames(p[[1]]) %in% "ENSDART00000166393"])
   # seq <- getSeq(genome, tx)
   # seq0 <- paste(as.character(seq), collapse="")
-  # seq01 <- unlist(lapply(re, function(.re) substr(seq0, start = .re, stop = .re+readsLen-1)))
+  # seq01 <- unlist(lapply(re, function(.re)
+  #       substr(seq0, start = .re, stop = .re+readsLen-1)))
   # grn <- gl[[1]]
-  # seq2 <- as.character(mcols(grn[mcols(grn)$tx %in% "ENSDART00000166393"])$seq)
+  # seq2 <-as.character(mcols(grn[mcols(grn)$tx %in% "ENSDART00000166393"])$seq)
   # all(seq01==seq2)
+  # ENSDART00000164359, ENSDART00000159144
   # tx <- cds[cds$tx_name %in% "ENSDART00000159144"]
-  # re <- start(p[[1]][seqnames(p[[1]]) %in% "ENSDART00000159144"]) #ENSDART00000164359
+  # re <- start(p[[1]][seqnames(p[[1]]) %in% "ENSDART00000159144"])
   # seq <- getSeq(genome, tx)
   # seq0 <- paste(as.character(seq), collapse="")
-  # seq01 <- unlist(lapply(re, function(.re) substr(seq0, start = .re, stop = .re+readsLen-1)))
-  # seq2 <- as.character(mcols(grn[mcols(grn)$tx %in% "ENSDART00000159144"])$seq)
+  # seq01 <- unlist(lapply(re, function(.re)
+  #          substr(seq0, start = .re, stop = .re+readsLen-1)))
+  # seq2 <-as.character(mcols(grn[mcols(grn)$tx %in% "ENSDART00000159144"])$seq)
   # all(seq01==seq2)
   if(!missing(outPath)){
     dir.create(outPath)